+++ apexpy/tests/test_fortranapex.py
@@ -21,7 +21,7 @@ from apexpy import fortranapex as fa
 
 
 class TestFortranApex():
-    def setup(self):
+    def setup_method(self):
         """Initialize each test."""
         fa.loadapxsh(os.path.join(os.path.dirname(apexpy.__file__),
                                   'apexsh.dat'), 2000)
@@ -55,7 +55,7 @@ class TestFortranApex():
         self.out = None
         self.test_out = None
 
-    def teardown(self):
+    def teardown_method(self):
         """Clean environment after each test."""
         del self.lat, self.lon, self.height, self.refh, self.vecflg
         del self.qlat, self.qlon, self.mlat, self.mlon, self.f1, self.f2
@@ -63,8 +63,19 @@ class TestFortranApex():
         del self.e2, self.e3, self.precision, self.glat, self.glon, self.error
         del self.out, self.test_out
 
-    def run_test_evaluation(self, rtol=1e-6, atol=1e-6):
-        """Run the evaluation of the test results."""
+    def run_test_evaluation(self, rtol=1e-5, atol=1e-5):
+        """Run the evaluation of the test results.
+
+        Parameters
+        ----------
+        rtol : float
+            Relative tolerance, default value based on old code's precision
+            (default=1e-5)
+        atol : float
+            Absolute tolerance, default value based on old code's precision
+            (default=1e-5)
+
+        """
 
         assert self.out is not None, "No results to test"
         assert self.test_out is not None, "No 'truth' results provided"